# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "screen-service API",
        "title": "screen-service API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/swagger/screen-service",
    "paths": {
...
        "/group/all-trees": {
            "get": {
                "description": "分组树",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Group"
                ],
                "summary": "分组树",
                "parameters": [
                    {
                        "type": "string",
                        "description": "分类",
                        "name": "type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "实例的数组",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/entity.GroupVo"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "错误code和错误信息",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
        "/group/batch-delete": {
            "post": {
                "description": "批量删除",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Group"
                ],
                "summary": "批量删除",
                "parameters": [
                    {
                        "description": "id数组",
                        "name": "ids",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功code和成功信息",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "500": {
                        "description": "错误code和错误信息",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
        "/group/batch-upsert": {
            "post": {
                "description": "批量更新",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Group"
                ],
                "summary": "批量更新",
                "parameters": [
                    {
                        "description": "对象数组",
                        "name": "entities",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "additionalProperties": true
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功code和成功信息",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "500": {
                        "description": "错误code和错误信息",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
        "/group/save": {
            "post": {
                "description": "保存分组",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Group"
                ],
                "summary": "保存分组",
                "parameters": [
                    {
                        "description": "Group全部信息",
                        "name": "item",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.GroupVoItem"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "实例",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/entity.GroupVoItem"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "错误code和错误信息",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
        "/group/{id}": {
            "delete": {
                "description": "删除实例",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Group"
                ],
                "summary": "删除实例",
                "parameters": [
                    {
                        "type": "string",
                        "description": "实例id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "实例",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.Group"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "错误code和错误信息",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "common.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "msg": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "entity.GroupVo": {
            "type": "object",
            "properties": {
                "children": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.GroupVo"
                    }
                },
                "id": {
                    "description": "id",
                    "type": "string"
                },
                "level": {
                    "description": "level",
                    "type": "integer"
                },
                "name": {
                    "description": "名称",
                    "type": "string"
                },
                "parentId": {
                    "description": "parent_id",
                    "type": "string"
                },
                "sortNum": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "entity.GroupVoItem": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "id",
                    "type": "string"
                },
                "level": {
                    "description": "level",
                    "type": "integer"
                },
                "name": {
                    "description": "名称",
                    "type": "string"
                },
                "parentId": {
                    "description": "parent_id",
                    "type": "string"
                },
                "sortNum": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "model.Group": {
            "type": "object",
            "properties": {
                "created_by": {
                    "description": "创建人",
                    "type": "string"
                },
                "created_time": {
                    "description": "创建时间",
                    "type": "string"
                },
                "id": {
                    "description": "唯一标识",
                    "type": "string"
                },
                "level": {
                    "description": "层次",
                    "type": "integer"
                },
                "name": {
                    "description": "名称",
                    "type": "string"
                },
                "parent_id": {
                    "description": "父id",
                    "type": "string"
                },
                "sort_num": {
                    "description": "排序",
                    "type": "integer"
                },
                "type": {
                    "description": "分类",
                    "type": "string"
                },
                "updated_by": {
                    "description": "更新人",
                    "type": "string"
                },
                "updated_time": {
                    "description": "更新时间",
                    "type": "string"
                }
            }
        },
...
    }
}